% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coverage_length.R
\name{compute_coverage_length}
\alias{compute_coverage_length}
\title{Compute coverage and length}
\usage{
compute_coverage_length(intervals)
}
\arguments{
\item{intervals}{List of individual and simultaneous intervals from simulations}
}
\value{
List with following parameters:
\item{cov_ind}{Empircial coverage of individual intervals}
\item{average_length_ind}{Average length of individual intervals}
\item{var_length_ind}{Variance of length of individual intervals}
\item{cov_sim}{Empircial coverage of simultaneous intervals}
\item{average_length_sim}{Average length of simultaneous intervals}
\item{var_length_sim}{Variance of length of simultaneous intervals}
}
\description{
Function to calculate the empirical coverage, the average length and the variance
of lengths of individual and simultaneous intervals
over simulation runs
}
