% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coverage_length.R
\name{empirical_coverage}
\alias{empirical_coverage}
\title{Compute empirical coverage}
\usage{
empirical_coverage(intervals_one_sample)
}
\arguments{
\item{intervals_one_sample}{List of individual and simultaneous intervals from
one simulation simulation}
}
\value{
List with following parameters:
\item{test_ind}{Empircial coverage of individual intervals}
\item{test_sim}{Empircial coverage of simultaneous intervals}
}
\description{
Function to calculate the empirical coverage of individual and simultaneous intervals
over simulation runs
}
