% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NERM.R
\name{fit_NERM}
\alias{fit_NERM}
\title{Fit NERM}
\usage{
fit_NERM(formula_y, data_sample, id_cluster, cluster_means)
}
\arguments{
\item{formula_y}{Formula for fitting NERM}

\item{data_sample}{Data frame}

\item{id_cluster}{Vector with cluster labels}

\item{cluster_means}{Cluster-level covariates for fixed parameters}
}
\value{
List with following parameters:
\item{var_u}{Estimated variance of random effects}
\item{var_e}{Estimated variance of errors}
\item{beta_hat}{Estimated regression parameters}
\item{u_hat}{Estimated random effects}
\item{mu_hat}{Estimated mixed effects}
\item{e_hat}{Estimated errors}
\item{mse}{Analytical mse}
\item{g1}{Analytical variance of mixed effect}
}
\description{
Function to fit nested error regression model
}
\examples{

set.seed(11992)
m = 25
n_j = 5
n = m * n_j
scaling_factor_e = 1
scaling_factor_u = sqrt(0.5)
x_ij = runif(n, 0, 1)
X = cbind(X0 = rep(1, n), X1 = x_ij)
t_X  = t(X)
id_cluster = rep(1:m, each = n_j)
beta = c(1, 1)

cluster_means <- as.matrix(aggregate(X,
list(id_cluster),
FUN = mean)[, -1])

#Formula to construct intervals
formula_y <-  y ~ 1 + X1 + (1| id_cluster)

data_sample <- generate_NERM(generate_u = list(type = "chisquared",
                                               scaling_factor = 1,
                                               dg = 6),
                             generate_e = list(type = "chisquared",
                                               scaling_factor = 1,
                                               dg = 6),
                             beta = beta,
                             X = X,
                             id_cluster = id_cluster,
                             start_seed = 1,
                             no_sim = 1,
                             cluster_means)


fitted_NERM <- fit_NERM(formula_y, data_sample,
                        id_cluster, cluster_means)

}
