% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_re.R
\name{generate_re}
\alias{generate_re}
\title{Generate centered and scaled random elements}
\usage{
generate_re(n, dg, scaling_factor, type = c("chisquare", "student_t"))
}
\arguments{
\item{n}{Number of observations}

\item{dg}{Degrees of freedom}

\item{scaling_factor}{Scaling factor}

\item{type}{Type of random element, \code{chisquare}: chi-squared distributed,
\code{student_t}: Student's t-distributed}
}
\value{
\item{random_elements}{Vector of random elements}
}
\description{
Generate centered and scaled random elements
}
\examples{

# Generate scaled and centered Student's t-distributed random elements

t_student_re <- generate_re(n = 100, dg = 6, scaling_factor = 1,
type = "student_t")


}
