% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_re.R
\name{generate_re}
\alias{generate_re}
\title{Generate random elements}
\usage{
generate_re(
  n,
  dg = 5,
  scaling_factor,
  type = c("chisquared", "student_t", "normal"),
  start_seed = 1
)
}
\arguments{
\item{n}{Number of observations}

\item{dg}{Degrees of freedom}

\item{scaling_factor}{Scaling factor}

\item{type}{Distribution of random element, \code{chisquare}: chi-squared distribution,
\code{student_t}: Student's t-distribution, \code{normal}: normal distribution}

\item{start_seed}{Seed to reproduce simulations}
}
\value{
\item{random_elements}{Vector of random elements}
}
\description{
Function to generate centered and scaled random effects or errors from chi-squared,
Student's t or normal distributions
}
\examples{

# Generate scaled and centered Student's t-distributed random elements

t_student_re <- generate_re(n = 100, dg = 6, scaling_factor = 1,
type = "student_t")


}
